+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for Project Utilities
 * Added `ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Posts\GetMetaDataInterface`.
 * Added `ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Posts\RegisterCustomTypeInterface`.
 * Added `ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Roles\ModifyUserRoleInterface`.
+* Added `ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Settings\SettingsRepositoryInterface`.
 * Added `ObsidianMoon\ProjectUtilities\Traits\WordPress\AddJFIFMimeTypeTrait`.
 * Added additional documentation for various sections of the code.